refactor: remove old ImageListForm and replaced with compose varient
