+++ app/src/androidMain/kotlin/de/westnordost/streetcomplete/quests/barrier_bicycle_barrier_installation/AddBicycleBarrierInstallationForm.kt
@@ -1,11 +1,11 @@
 package de.westnordost.streetcomplete.quests.barrier_bicycle_barrier_installation
 
 import de.westnordost.streetcomplete.R
-import de.westnordost.streetcomplete.quests.AImageListQuestComposeForm
+import de.westnordost.streetcomplete.quests.AImageListQuestForm
 import de.westnordost.streetcomplete.quests.AnswerItem
 
 class AddBicycleBarrierInstallationForm :
-    AImageListQuestComposeForm<BicycleBarrierInstallation, BicycleBarrierInstallationAnswer>() {
+    AImageListQuestForm<BicycleBarrierInstallation, BicycleBarrierInstallationAnswer>() {
 
     override val items = BicycleBarrierInstallation.entries.map { it.asItem() }
     override val itemsPerRow = 3
