+++ app/src/androidMain/kotlin/de/westnordost/streetcomplete/quests/bollard_type/AddBollardTypeForm.kt
@@ -1,10 +1,10 @@
 package de.westnordost.streetcomplete.quests.bollard_type
 
 import de.westnordost.streetcomplete.R
-import de.westnordost.streetcomplete.quests.AImageListQuestComposeForm
+import de.westnordost.streetcomplete.quests.AImageListQuestForm
 import de.westnordost.streetcomplete.quests.AnswerItem
 
-class AddBollardTypeForm : AImageListQuestComposeForm<BollardType, BollardTypeAnswer>() {
+class AddBollardTypeForm : AImageListQuestForm<BollardType, BollardTypeAnswer>() {
 
     override val items = BollardType.entries.map { it.asItem() }
     override val itemsPerRow = 3
