+++ app/src/androidMain/kotlin/de/westnordost/streetcomplete/quests/drinking_water_type/AddDrinkingWaterTypeForm.kt
@@ -1,8 +1,8 @@
 package de.westnordost.streetcomplete.quests.drinking_water_type
 
-import de.westnordost.streetcomplete.quests.AImageListQuestComposeForm
+import de.westnordost.streetcomplete.quests.AImageListQuestForm
 
-class AddDrinkingWaterTypeForm : AImageListQuestComposeForm<DrinkingWaterType, DrinkingWaterType>() {
+class AddDrinkingWaterTypeForm : AImageListQuestForm<DrinkingWaterType, DrinkingWaterType>() {
 
     override val items = DrinkingWaterType.entries.map { it.asItem() }
