+++ app/src/androidMain/kotlin/de/westnordost/streetcomplete/quests/fire_hydrant/AddFireHydrantTypeForm.kt
@@ -1,8 +1,8 @@
 package de.westnordost.streetcomplete.quests.fire_hydrant
 
-import de.westnordost.streetcomplete.quests.AImageListQuestComposeForm
+import de.westnordost.streetcomplete.quests.AImageListQuestForm
 
-class AddFireHydrantTypeForm : AImageListQuestComposeForm<FireHydrantType, FireHydrantType>() {
+class AddFireHydrantTypeForm : AImageListQuestForm<FireHydrantType, FireHydrantType>() {
 
     override val items = FireHydrantType.entries.map { it.asItem() }
     override val itemsPerRow = 2
