+++ app/src/androidMain/kotlin/de/westnordost/streetcomplete/quests/leaf_detail/AddTreeLeafTypeForm.kt
@@ -1,11 +1,11 @@
 package de.westnordost.streetcomplete.quests.leaf_detail
 
 import de.westnordost.streetcomplete.R
-import de.westnordost.streetcomplete.quests.AImageListQuestComposeForm
+import de.westnordost.streetcomplete.quests.AImageListQuestForm
 import de.westnordost.streetcomplete.quests.AnswerItem
 
 class AddTreeLeafTypeForm :
-    AImageListQuestComposeForm<TreeLeafType, TreeLeafTypeAnswer>() {
+    AImageListQuestForm<TreeLeafType, TreeLeafTypeAnswer>() {
 
     override val items = TreeLeafType.entries.map { it.asItem() }
     override val itemsPerRow = 2
