+++ app/src/androidMain/kotlin/de/westnordost/streetcomplete/quests/oneway/AddOnewayForm.kt
@@ -5,12 +5,12 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.key
 import androidx.compose.ui.semantics.Role
 import de.westnordost.streetcomplete.data.osm.geometry.ElementPolylinesGeometry
-import de.westnordost.streetcomplete.quests.AImageListQuestComposeForm
+import de.westnordost.streetcomplete.quests.AImageListQuestForm
 import de.westnordost.streetcomplete.ui.common.image_select.ImageListItem
 import de.westnordost.streetcomplete.ui.common.image_select.SelectableIconItem
 import de.westnordost.streetcomplete.util.math.getOrientationAtCenterLineInDegrees
 
-class AddOnewayForm : AImageListQuestComposeForm<OnewayAnswer, OnewayAnswer>() {
+class AddOnewayForm : AImageListQuestForm<OnewayAnswer, OnewayAnswer>() {
 
     override val items get() =
         OnewayAnswer.entries.map { it.asItem(requireContext(), wayRotation - mapRotation) }
