+++ app/src/androidMain/kotlin/de/westnordost/streetcomplete/quests/orchard_produce/AddOrchardProduceForm.kt
@@ -1,8 +1,8 @@
 package de.westnordost.streetcomplete.quests.orchard_produce
 
-import de.westnordost.streetcomplete.quests.AImageListQuestComposeForm
+import de.westnordost.streetcomplete.quests.AImageListQuestForm
 
-class AddOrchardProduceForm : AImageListQuestComposeForm<OrchardProduce, List<OrchardProduce>>() {
+class AddOrchardProduceForm : AImageListQuestForm<OrchardProduce, List<OrchardProduce>>() {
 
     private val producesMap = OrchardProduce.entries
         .map { it.asItem() }
