+++ app/src/androidMain/kotlin/de/westnordost/streetcomplete/quests/powerpoles_material/AddPowerPolesMaterialForm.kt
@@ -1,10 +1,10 @@
 package de.westnordost.streetcomplete.quests.powerpoles_material
 
 import de.westnordost.streetcomplete.R
-import de.westnordost.streetcomplete.quests.AImageListQuestComposeForm
+import de.westnordost.streetcomplete.quests.AImageListQuestForm
 import de.westnordost.streetcomplete.quests.AnswerItem
 
-class AddPowerPolesMaterialForm : AImageListQuestComposeForm<PowerPolesMaterial, PowerPolesMaterialAnswer>() {
+class AddPowerPolesMaterialForm : AImageListQuestForm<PowerPolesMaterial, PowerPolesMaterialAnswer>() {
 
     override val items = PowerPolesMaterial.entries.map { it.asItem() }
     override val itemsPerRow = 3
