+++ app/src/androidMain/kotlin/de/westnordost/streetcomplete/quests/religion/AddReligionForm.kt
@@ -5,13 +5,13 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.key
 import androidx.compose.ui.semantics.Role
 import de.westnordost.streetcomplete.R
-import de.westnordost.streetcomplete.quests.AImageListQuestComposeForm
+import de.westnordost.streetcomplete.quests.AImageListQuestForm
 import de.westnordost.streetcomplete.quests.AnswerItem
 import de.westnordost.streetcomplete.quests.religion.Religion.MULTIFAITH
 import de.westnordost.streetcomplete.ui.common.image_select.ImageListItem
 import de.westnordost.streetcomplete.ui.common.image_select.SelectableIconItem
 
-class AddReligionForm : AImageListQuestComposeForm<Religion, Religion>() {
+class AddReligionForm : AImageListQuestForm<Religion, Religion>() {
 
     override val itemContent =
         @Composable { item: ImageListItem<Religion>, index: Int, onClick: () -> Unit, role: Role ->
