+++ app/src/androidMain/kotlin/de/westnordost/streetcomplete/quests/surface/AddPathSurfaceForm.kt
@@ -4,11 +4,11 @@ import de.westnordost.streetcomplete.R
 import de.westnordost.streetcomplete.data.osm.mapdata.Way
 import de.westnordost.streetcomplete.osm.surface.Surface
 import de.westnordost.streetcomplete.osm.surface.toItems
-import de.westnordost.streetcomplete.quests.AImageListQuestComposeForm
+import de.westnordost.streetcomplete.quests.AImageListQuestForm
 import de.westnordost.streetcomplete.quests.AnswerItem
 import de.westnordost.streetcomplete.util.ktx.couldBeSteps
 
-class AddPathSurfaceForm : AImageListQuestComposeForm<Surface, SurfaceOrIsStepsAnswer>() {
+class AddPathSurfaceForm : AImageListQuestForm<Surface, SurfaceOrIsStepsAnswer>() {
     override val items get() = Surface.selectableValuesForWays.toItems()
 
     override val otherAnswers get() = listOfNotNull(
